+++ cgbeacon2/server/blueprints/api_v1/controllers.py
@@ -201,7 +201,7 @@ def create_allele_query(resp_obj, req):
     if "includeDatasetResponses" not in customer_query:
         customer_query["includeDatasetResponses"] = "NONE"
 
-    # check if the minimal required params were provided in query
+    # check if the minimum required params were provided in query
     check_allele_request(resp_obj, customer_query, mongo_query)
 
     # if an error occurred, do not query database and return error
@@ -215,6 +215,25 @@ def create_allele_query(resp_obj, req):
     return mongo_query
 
 
+def check_request_build(datasets, build):
+    """Make sure that genome build requested in query is available on this beacon for the given datasets.
+
+    Accepts:
+        datasets(list): list of datasets present in the request
+        build(str): genome build of variant present in the request
+
+    Returns:
+        True or False
+    """
+    ds_query = {}
+    if len(datasets) > 0:
+        ds_query["_id"] = {"$in": datasets}
+
+    results = current_app.db["dataset"].find(ds_query, {"assembly_id": 1, "_id": 0})
+    ds_builds = [res["assembly_id"] for res in results]
+    return build in ds_builds
+
+
 def check_allele_request(resp_obj, customer_query, mongo_query):
     """Check that the query to the server is valid
 
@@ -223,34 +242,38 @@ def check_allele_request(resp_obj, customer_query, mongo_query):
         customer_query(dict): a dictionary with all the key/values provided in the external request
         mongo_query(dict): the query to collect variants from this server
     """
-    # If customer asks for a classical SNV
-    if customer_query.get("variantType") is None and all(
-        [
-            customer_query.get("referenceName"),
-            customer_query.get(
-                "start",
-            ),
-            customer_query.get("end"),
-            customer_query.get("referenceBases"),
-            customer_query.get("alternateBases"),
-            customer_query.get("assemblyId"),
-        ]
+    chrom = customer_query.get("referenceName")
+    start = customer_query.get("start")
+    end = customer_query.get("end")
+    ref = customer_query.get("referenceBases")
+    alt = customer_query.get("alternateBases")
+    build = customer_query.get("assemblyId")
+    datasets = customer_query.get("datasetIds", [])
+    variant_type = customer_query.get("variantType")
+
+    # If customer wants to match a SNV with precise coordinates, alt and ref
+    if (
+        customer_query.get("variantType") is None
+        and all([chrom, start, end, ref, alt, build])
+        and not "N" in ref
+        and not "N" in alt
     ):
-        # generate md5_key to compare with our database
+        # generate md5_key to quickly compare with our database
         mongo_query["_id"] = md5_key(
-            customer_query["referenceName"],
-            customer_query["start"],
-            customer_query.get("end"),
-            customer_query["referenceBases"],
-            customer_query["alternateBases"],
-            customer_query["assemblyId"],
+            chrom,
+            start,
+            end,
+            ref,
+            alt,
+            build,
         )
+        return
 
     # Check that the 3 mandatory parameters are present in the query
     if None in [
-        customer_query.get("referenceName"),
-        customer_query.get("referenceBases"),
-        customer_query.get("assemblyId"),
+        chrom,
+        ref,
+        build,
     ]:
         # return a bad request 400 error with explanation message
         resp_obj["message"] = dict(
@@ -259,27 +282,19 @@ def check_allele_request(resp_obj, customer_query, mongo_query):
         )
         return
 
-    # check if genome build requested corresponds to genome build of the available datasets:
-    if len(customer_query.get("datasetIds", [])) > 0:
-        dset_builds = current_app.db["dataset"].find(
-            {"_id": {"$in": customer_query["datasetIds"]}}, {"assembly_id": 1, "_id": 0}
+    if len(datasets) > 0 and check_request_build(datasets, build) is False:
+        resp_obj["message"] = dict(
+            error=BUILD_MISMATCH,
+            allelRequest=customer_query,
         )
-        dset_builds = [dset["assembly_id"] for dset in dset_builds if dset["assembly_id"]]
-        for dset in dset_builds:
-            if dset != customer_query["assemblyId"]:
-                # return a bad request 400 error with explanation message
-                resp_obj["message"] = dict(
-                    error=BUILD_MISMATCH,
-                    allelRequest=customer_query,
-                )
-                return
+        return
 
     # alternateBases OR variantType is also required
     if all(
         param is None
         for param in [
-            customer_query.get("alternateBases"),
-            customer_query.get("variantType"),
+            alt,
+            variant_type,
         ]
     ):
         # return a bad request 400 error with explanation message
@@ -288,9 +303,10 @@ def check_allele_request(resp_obj, customer_query, mongo_query):
             allelRequest=customer_query,
         )
         return
+
     # Check that genomic coordinates are provided (even rough)
     if (
-        customer_query.get("start") is None
+        start is None
         and any([coord in customer_query.keys() for coord in RANGE_COORDINATES]) is False
     ):
         # return a bad request 400 error with explanation message
@@ -300,12 +316,11 @@ def check_allele_request(resp_obj, customer_query, mongo_query):
         )
         return
 
-    if customer_query.get("start"):  # query for exact position
+    if start:  # query for exact position
         try:
-            if customer_query.get("end") is not None:
-                mongo_query["end"] = int(customer_query["end"])
-
-            mongo_query["start"] = int(customer_query["start"])
+            if end is not None:
+                mongo_query["end"] = int(end)
+            mongo_query["start"] = int(start)
 
         except ValueError:
             # return a bad request 400 error with explanation message
@@ -341,22 +356,31 @@ def check_allele_request(resp_obj, customer_query, mongo_query):
         if fuzzy_end_query:
             mongo_query["end"] = fuzzy_end_query
 
-    if mongo_query.get("_id") is None:
-        # perform normal query
-        mongo_query["assemblyId"] = customer_query["assemblyId"]
-        mongo_query["referenceName"] = customer_query["referenceName"]
-        mongo_query["referenceBases"] = customer_query["referenceBases"]
+    mongo_query["assemblyId"] = build
+    mongo_query["referenceName"] = chrom
 
-        if "alternateBases" in customer_query:
-            mongo_query["alternateBases"] = customer_query["alternateBases"]
+    add_coords_query(mongo_query, "referenceBases", ref)
 
-        if "variantType" in customer_query:
-            mongo_query["variantType"] = customer_query["variantType"]
+    if "alternateBases" in customer_query:
+        add_coords_query(mongo_query, "alternateBases", alt)
 
+    if "variantType" in customer_query:
+        mongo_query["variantType"] = variant_type
+
+
+def add_coords_query(mongo_query, field, value):
+    """Created a regex for a database query when ref or alt coords contain Ns
+
+    Accepts:
+        mongo_query(dict): an allele query dictionary
+        field(string): "referenceBases" or "alternateBases"
+        value(string): A stretch of bases, might containg Ns
+    """
+
+    if "N" in value:
+        mongo_query[field] = {"$regex": value.replace("N", ".")}
     else:
-        # use only variant _id in query
-        mongo_query.pop("start")
-        mongo_query.pop("end", None)
+        mongo_query[field] = value
 
 
 def dispatch_query(mongo_query, response_type, datasets=[], auth_levels=([], False)):